+++ src/shared/components/main/components/board/Boardgame.tsx
