+++ src/shared/components/main/components/stampBoard/Progress.tsx
