+++ src/shared/components/main/components/stampBoard/Stamp.tsx
