+++ src/shared/components/main/components/stampBoard/StampBoard.tsx
