+++ src/shared/components/main/components/video/VideoPlayer.tsx
